requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langsmith
faiss-cpu
sentence-transformers
pypdf
python-dotenv
easyocr
Pillow
duckduckgo-search